Fix inverted condition on year starting in DST
The first transition in a year with DST is *to* DST precisely if the
year *doesn't* start in DST, not if it does.

Also fix the startsInDst() check it's using there, which neglected the
possibility that the year's DST transition might be faked.

This amends commit d98b43005eed051a77beacb252c7a9eb7f5c4493

Pick-to: 6.3
Change-Id: I17422fa9972f826b5bf7e325f5e579f41300fd97
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>