Backport recent QFlags tests
This is a backport of the tst_qflags parts of

- 6f50e54138f9a2aa9ab6d84add449f197d545de3
  (QFlags: fix missing bitwise xor operators in QT_TYPESAFE_FLAGS builds)
- 53890b5d3cd771d66d6b73940eb38e3c3fc6d60f
  (qtestmouse.h: compile with QT_TYPESAFE_FLAGS)

These changes were originally only picked to 6.3 because it was
thought only 6.3 was failing the tests. In order to check that this is
indeed the case, and to help future cherry-picking, backport the
tst_qflags additions to 6.2 and 5.15.

Change-Id: Iecfa11a8a83388c60f659238a28a26142b579380
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>
(cherry picked from commit 4feaf095f44861fb93694483ac209b14974bf037)
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>