Fix high-dpi and QScreen state initialization
Fix bug where QScreen::geometry() returns incorrect
(unscaled) geometry.

We maintain state:

  - QHighDpiScaling::m_active, which is set to true
    if any screen in QGuiApplication::screens() has a
    scale factor > 1, in updateHighDpiScaling()
  - QScreen::m_geometry, which is assigned the scaled
    screen size in QScreenPrivate::updateHighDpi()

Now the problem is that updateHighDpiScaling() requires
an up-to date screens list, and that updateHighDpi()
requires an up-to date m_active, and that currently
updateHighDpi() is called when constructing the QScreen
only.

This is all a bit unsatisfactory, but for now fix this
in the simplest way by adding a second call to
updateHighDpi() in handleScreenAdded(), after the call
to updateHighDpiScaling().

Fixes-number: QTBUG-101601
Change-Id: I5f442ceb966c77d6746ee885203b67da96612587
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 942de89562f363afcd4a399dcd55c64792fd5b7c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>